+++ ImageFeed/Models/Photo.swift
@@ -1,0 +1,46 @@
+//
+//  Photo.swift
+//  ImageFeed
+//
+//  Created by Main on 08.05.2025.
+//
+
+import Foundation
+
+struct Photo {
+    let id: String
+    let size: CGSize
+    let createdAt: Date?
+    let welcomeDescription: String?
+    let thumbImageURL: String
+    let largeImageURL: String
+    let isLiked: Bool
+}
+
+struct PhotoResult: Decodable {
+    let id: String
+    let width: Double
+    let height: Double
+    let createdAt: String?
+    let description: String?
+    let urls: UrlsResult
+    let likedByUser: Bool
+    
+    enum CodingKeys: String, CodingKey {
+        case id
+        case width
+        case height
+        case createdAt = "created_at"
+        case description
+        case urls
+        case likedByUser = "liked_by_user"
+    }
+}
+
+struct UrlsResult: Codable {
+    let raw: String
+    let full: String
+    let regular: String
+    let small: String
+    let thumb: String
+}
